Move code to a src folder
