+++ plugins/index.ts
@@ -3,7 +3,7 @@ import { mermaidPlugin } from "./mermaid";
 import { sillyPlugin } from "./silly-effects";
 import { codeDetectionPlugin } from "./code-detection";
 import { markdownLogging } from "./markdown-logging";
-import {stackSnippetExternal} from "./stack-snippets/stack-snippet-external";
+import {stackSnippetPlugin} from "./stack-snippets/src/stackSnippetPlugin";
 
 /**
  * Plugins written to demonstrate potential capabilities of plugins.
@@ -20,7 +20,7 @@ export const samplePlugins = [
  * Plugins written and maintained by Stack developers, intended for production deployment.
  */
 export const firstPartyPlugins = [
-    stackSnippetExternal
+    stackSnippetPlugin
 ];
 
 /**
