+++ plugins/stack-snippets/src/paste-handler.ts
@@ -1,7 +1,7 @@
 import { Plugin, NodeSelection } from "prosemirror-state";
 import { EditorView } from "prosemirror-view";
 import { Node as ProsemirrorNode, Schema, Slice } from "prosemirror-model";
-import { log } from "../../src/shared/logger";
+import { log } from "../../../src/shared/logger";
 import {
     BeginMetaLine,
     EndMetaLine,
@@ -9,8 +9,8 @@ import {
     mapMetaLine,
     validSnippetRegex,
 } from "./common";
-import { generateRandomId } from "../../src/shared/utils";
-import { insertParagraphIfAtDocEnd } from "../../src/rich-text/commands/helpers";
+import { generateRandomId } from "../../../src/shared/utils";
+import { insertParagraphIfAtDocEnd } from "../../../src/rich-text/commands/helpers";
 
 export const parseSnippetBlockForProsemirror = (
     schema: Schema,
